Added pre-setup and progress on Git
Added some notes for setting up a Mac for development, specifically XCode tools and homebrew
Covered the absolute basics of getting a Git repository on Github. Still need to cover the common workflow. Going to rely on external tutorials for branching and whatnot, since visualization and such is helpful for that.